requirements.txt
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==4.2.11
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.4
pillow==10.3.0
proglog==0.1.10
python-dotenv==1.0.1
pytube==15.0.0
redis==5.0.3
requests==2.31.0
six==1.16.0
sniffio==1.3.1
spotipy==2.23.0
sqlparse==0.4.4
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
youtube-search-python==1.6.6
